Data for Songs on Taylor's Official Studio Albums

A data set containing lyrics to and characteristics of songs on all of Taylor's official studio albums. Thus, this is a subset of tswift_all_songs, with EPs and individual singles excluded. Critically, this subset also only includes versions owned by Taylor when possible, because we stan artists owning their own work. This means that although both Fearless and Fearless (Taylor's Version) appear in full tswift_all_songs data, only Fearless (Taylor's Version) appears in this subset. This also means that this data set will change as additional re-releases are made available (i.e., Red will soon be replaced with Red (Taylor's Version)).

tswift_album_songs

Format

This is a subset of the tswift_all_songs data set. Please see that documentation for a complete description of all the included fields.