Color palettes based on Taylor Swift's albums

Premade color palettes based on Taylor Swifts album covers. For details on how to extend and shorten these palettes, or create your own color palette, see color_palette().

Usage

album_palettes

album_compare

Format

A list of length 11. Each element contains a color palette for one of Taylor Swift's studio albums. The list elements are named according to the name of the album.

An object of class taylor_color_palette (inherits from vctrs_vctr, character) of length 11.

Source

Colors derived from album covers using 'Colormind'.

See also

Examples

album_palettes
#> $taylor_swift
#> <color_palette[5]>
#>     #1D4737 
#>     #81A757 
#>     #1BAEC6 
#>     #523d28 
#>     #E7DBCC 
#> 
#> $fearless
#> <color_palette[5]>
#>     #6E4823 
#>     #976F34 
#>     #CBA863 
#>     #ECD59F 
#>     #E1D4C2 
#> 
#> $fearless_tv
#> <color_palette[5]>
#>     #624324 
#>     #A47F45 
#>     #CAA462 
#>     #C5AA7C 
#>     #EEDBA9 
#> 
#> $speak_now
#> <color_palette[5]>
#>     #2E1924 
#>     #6C3127 
#>     #833C63 
#>     #D1A0C7 
#>     #F5E8E2 
#> 
#> $red
#> <color_palette[5]>
#>     #201F39 
#>     #A91E47 
#>     #7E6358 
#>     #B0A49A 
#>     #DDD8C9 
#> 
#> $red_tv
#> <color_palette[5]>
#>     #400303 
#>     #731803 
#>     #967862 
#>     #B38468 
#>     #C7C5B6 
#> 
#> $`1989`
#> <color_palette[5]>
#>     #5D4E5D 
#>     #846578 
#>     #92573C 
#>     #C6B69C 
#>     #D8D8CF 
#> 
#> $reputation
#> <color_palette[5]>
#>     #2C2C2C 
#>     #515151 
#>     #5B5B5B 
#>     #6E6E6E 
#>     #B9B9B9 
#> 
#> $lover
#> <color_palette[5]>
#>     #76BAE0 
#>     #8C4F66 
#>     #B8396B 
#>     #EBBED3 
#>     #FFF5CC 
#> 
#> $folklore
#> <color_palette[5]>
#>     #3E3E3E 
#>     #545454 
#>     #5C5C5C 
#>     #949494 
#>     #EBEBEB 
#> 
#> $evermore
#> <color_palette[5]>
#>     #160E10 
#>     #421E18 
#>     #D37F55 
#>     #85796D 
#>     #E0D9D7 
#> 

album_compare
#> <color_palette[11]>
#>     taylor_swift 
#>     fearless 
#>     fearless_tv 
#>     speak_now 
#>     red 
#>     red_tv 
#>     1989 
#>     reputation 
#>     lover 
#>     folklore 
#>     evermore 

album_palettes$evermore
#> <color_palette[5]>
#>     #160E10 
#>     #421E18 
#>     #D37F55 
#>     #85796D 
#>     #E0D9D7